+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This version corresponds to Magba v0.4.3.
 ### Bug Fixes
 
 - Fix collection constructors interpreting children pose as in local frame. Now assume children poses in global frame.
+- Fix stub return types from `pymagba_binding` classes to `pymagba` classes.
 - Use `ArrayLike` in stubs instead of Sequence.
 - Fix default polarization vectors of magnet classes to [0, 0, 1].